requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
dataclasses-json==0.5.7
faiss-cpu==1.7.3
filelock==3.9.0
frozenlist==1.3.3
idna==3.4
langchain==0.0.85
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.26.5
packaging==23.0
pycryptodomex==3.17
pydantic==1.10.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
tenacity==8.2.1
tiktoken==0.2.0
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
